my_docker_project/bot_app/requirements.txt
flask
gunicorn
requests
line-bot-sdk